Finally decided how to handle common UI states
Using namespaced classes for states such as focus, active, disabled, selected, hover is very verbose. So i've introduced a global namespace for these special states (they're not really modifiers because they're run-time changes rather than design time), so we have 'mediacat-is-active', 'mediacat-is-disabled' and so on. Hopefully these are the only exception to the no-multi-class-selectors rule.